+++ app/main/__init__.py
@@ -1,0 +1,6 @@
+from flask import Blueprint
+
+main = Blueprint('main', __name__)
+
+from .views import *
+from .errors import *
