+++ app/main/models/user.py
@@ -1,0 +1,16 @@
+from app import db
+from .role import Role
+
+class User(db.Model):
+    __tablename__ = 'users'
+
+    id = db.Column(db.Integer, primary_key=True)
+    name = db.Column(db.String(16), index=True, unique=True)
+    role_id = db.Column(db.Integer, db.ForeignKey('roles.id'))
+
+
+    def __repr__(self):
+        return '<User {0}>'.format(self.name)
+
+
+
