+++ app/templates/base.html
@@ -17,11 +17,9 @@
         </div>
         <div class="navbar-collapse collapse">
             <ul class="nav navbar-nav">
-                <li><a href="{{ url_for('index') }}">Home</a></li>
-                <li><a href="{{ url_for('hello_user') }}">User</a></li>
-                <li><a href="{{ url_for('image') }}">Image</a></li>
-                <li><a href="{{ url_for('session_test') }}">Session</a></li>
-                <li><a href="{{ url_for('geekhub_redirect') }}">Geekhub</a></li>
+                <li><a href="{{ url_for('main.index') }}">Home</a></li>
+                <li><a href="{{ url_for('main.hello_user') }}">User</a></li>
+                <li><a href="{{ url_for('main.image') }}">Image</a></li>
             </ul>
         </div>
     </div>
