+++ app/templates/main/image.html
@@ -9,8 +9,8 @@
                 {{ wtf.quick_form(form, enctype='multipart/form-data') }}
             </div>
         </div>
-        {% if image and name%}
-            <h3>{{ name }}</h3>
+        {% if image%}
+            <h3>{{ image and name}}</h3>
             <img src="{{ url_for('static', filename=image) }}">
         {% endif %}
     </div>
