--- models/user.py
@@ -1,26 +1,0 @@
-from models import db
-
-class User(db.Model):
-    __tablename__ = 'user'
-
-    id = db.Column(db.Integer, primary_key=True)
-    name = db.Column(db.String(16), index=True, unique=True)
-    role_id = db.Column(db.Integer, db.ForeignKey('roles.id'))
-
-
-    def __repr__(self):
-        return '<User {0}'.format(self.name)
-
-class Role(db.Model):
-    __tablename__ = 'roles'
-    id = db.Column(db.Integer, primary_key=True)
-    name = db.Column(db.String(64), unique=True)
-    users = db.relationship('User', backref='role', lazy='dynamic')
-
-    def __repr__(self):
-        return '<Role %r>' % self.name
-
-admin_role = Role(name='Admin')
-user_role = Role(name='User')
-
-
