--- templates/session.html
@@ -1,8 +1,0 @@
-{% extends "base.html" %}
-
-{% block title %}GeekHub Image{% endblock %}
-
-{% block content %}
-<h1>You have seen this page {{ count }} times.</h1>
-<p>Hit refresh to increase.</p>
-{% endblock %}
